fix broken links in News page
