+++ src/components/News/NewsCard.jsx
@@ -11,7 +11,7 @@ import {
     HStack,
 } from '@chakra-ui/react';
 import { useStockNews } from '../../contexts/useStockNews';
-import NewsArticle from '../NewsArticle';
+import NewsArticle from './NewsArticle';
 import CustomBox from '../CustomBox';
 
 const NewsCard = ({ source, children, ...otherProps }) => {
