+++ src/hooks/useDateFormat.js
@@ -1,7 +1,7 @@
 const useDateFormat = () => (date) => {
     if (date) {
         const dateObj = new Date(date.split('T')[0]);
-        let options = { year: 'numeric', month: 'long', day: '2-digit' };
+        const options = { year: 'numeric', month: 'long', day: '2-digit' };
         return dateObj.toLocaleDateString('en-US', options);
     } else {
         return null;
